pyproject.toml
[tool.poetry]
name = "saas-metering-platform"
version = "0.1.0"
description = "Multi-Tenant Subscription & Usage Metering Platform"
authors = ["Murali Krishna Pendyala"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
argon2-cffi = "^23.1.0"
psycopg2-binary = "^2.9.9" # For synchronous checks if needed, but mainly focusing on async

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
httpx = "^0.26.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["backend"]
asyncio_mode = "auto"
